% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToDataFrame.R
\name{convertToDataFrame}
\alias{convertToDataFrame}
\title{Convert a XML into a dataframe and name the columns}
\usage{
convertToDataFrame(data, colnames)
}
\arguments{
\item{data}{The XML data read}

\item{colnames}{A vector with the column names (it has to be the equal to the number of the dataframe columns)}
}
\value{
A dataframe built with the parameters
}
\description{
Convert a XML into a dataframe and name the columns
}
\examples{
Basic conversion to a dataframe
cols <- c("Sales", "Development", "Human resources")
convertToDataFrame(extractAttachmentContent("pdfdocwithattach.pdf")), cols)
}
