% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transposeDataFrame.R
\name{transposeDataFrame}
\alias{transposeDataFrame}
\title{Transpose a dataframe and name the columns}
\usage{
transposeDataFrame(dataframe, colnames)
}
\arguments{
\item{dataframe}{The dataframe to be transposed}

\item{colnames}{A vector with the column names (it has to be the equal to the number of the dataframe columns)}
}
\value{
A transposed dataframe built with the parameters
}
\description{
Transpose a dataframe and name the columns
}
\examples{
Basic transposition of a dataframe
cols <- c("Sales", "Development", "Human resources")
transposeDataFrame(convertToDataFrame(extractAttachmentContent("pdfdocwithattach.pdf"))), cols)
}
