% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateXSD.R
\name{validateXSD}
\alias{validateXSD}
\title{Validate a XML against the PDFDriver XSD}
\usage{
validateXSD(xml)
}
\arguments{
\item{xml}{The XML data read}
}
\value{
A boolean value indicating whether XML is validated
}
\description{
Validate a XML against the PDFDriver XSD
}
\examples{
Basic validation
validateXSD(extractMetadataContent("pdfdoc.pdf", "breakfast_menu"))
}
